Disable the UART lines while in power-down mode
When the device is powered off but the UART is attached
the little bit of power from the UART line can keep
parts of the test device awake, making the next boot
act weird. This puts the UART pins in input mode
while in the poweroff state to prevent power leakage.